refactor(grpc): isolate graceful server shutdown
