+++ cmd/jobs-service/main.go
@@ -20,6 +20,7 @@ import (
 	"github.com/hitesh22rana/chronoverse/internal/pkg/auth"
 	"github.com/hitesh22rana/chronoverse/internal/pkg/clickhouse"
 	grpcclient "github.com/hitesh22rana/chronoverse/internal/pkg/grpc/client"
+	grpcserverpkg "github.com/hitesh22rana/chronoverse/internal/pkg/grpcserver"
 	loggerpkg "github.com/hitesh22rana/chronoverse/internal/pkg/logger"
 	"github.com/hitesh22rana/chronoverse/internal/pkg/meilisearch"
 	"github.com/hitesh22rana/chronoverse/internal/pkg/postgres"
@@ -203,7 +204,7 @@ func run() int {
 		return ExitError
 	}
 
-	go svcpkg.GracefulStop(ctx, app, 20*time.Second)
+	go grpcserverpkg.GracefulStop(ctx, app, 20*time.Second)
 
 	// Log the service information
 	loggerpkg.FromContext(ctx).Info(
